trekk ny begrunnelseLøsning i egen fil. Rename GrunnlagMedForrigePeriodeOgBehandling -> BegrunnelseGrunnlag
